+++ mols/mols.h
@@ -13,22 +13,22 @@ class MOLS
 {
 public:
     MOLS() {}
-    MOLS(std::vector<float> x, std::vector<float> y);
+    MOLS(std::vector<double> x, std::vector<double> y);
     void defW();
     void defY();
-    float defTimeLimit(float sizeLimit);
+    double defTimeLimit(double sizeLimit);
 
-    Matrix<float> getW() { return _w; }
-    Matrix<float> getX() { return _X; }   
-    Matrix<float> getY() { return _Y; }
+    Matrix<double> getW() { return _w; }
+    Matrix<double> getX() { return _X; }   
+    Matrix<double> getY() { return _Y; }
 
 
 private:
-    Matrix<float> _X;
-    Matrix<float> _Y;
-    Matrix<float> _w;
+    Matrix<double> _X;
+    Matrix<double> _Y;
+    Matrix<double> _w;
 
-    std::vector<float> _x;
-    std::vector<float> _y;
+    std::vector<double> _x;
+    std::vector<double> _y;
 };
